+++ src/node/desktop/src/main/application.ts
@@ -44,8 +44,9 @@ import {
   resolveAliasedPath
 } from './utils';
 import { WindowTracker } from './window-tracker';
-import { configureSatelliteWindow, configureSecondaryWindow, focusedWebContents } from './window-utils';
+import { configureSatelliteWindow, configureSecondaryWindow } from './window-utils';
 import { Client, Server } from 'net-ipc';
+import { setTimeoutPromise } from '../core/wait-utils';
 
 /**
  * The RStudio application
@@ -60,6 +61,7 @@ export class Application implements AppState {
   gwtCallback?: GwtCallback;
   sessionStartDelaySeconds = 0;
   sessionEarlyExitCode = 0;
+  startupDelay = 0;
   pendingWindows = new Array<PendingWindow>();
   server?: Server;
   client?: Client;
@@ -290,6 +292,10 @@ export class Application implements AppState {
     initializeLang();
 
     this.argsManager.handleAppReadyCommands(this);
+
+    if (this.startupDelay) {
+      await setTimeoutPromise(this.startupDelay * 1000);
+    }
 
     // on Windows, ask the user what version of R they'd like to use
     let rPath = '';